+++ core/database/tnlcm-structure.js
@@ -14,6 +14,7 @@ var tnlcmAdminEmail = process.env.TNLCM_ADMIN_EMAIL;
 var db = db.getSiblingDB(dbName);
 
 // Create collections
+db.createCollection("resource_manager");
 db.createCollection("trial_network");
 db.createCollection("user");